LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-scroll-position-try-006-expected.txt


PASS Should use the last (fourth) position option initially
FAIL Should still use the last position option as long as it fits. assert_equals: Anchored element should be at the left of anchor expected 650 but got 800
PASS Should use the third position option with enough space below, and not enough above
FAIL Should use the second position option with enough space right assert_equals: Anchored element should be at the right of anchor expected 650 but got 550
FAIL Should still use the second position option as long as it fits assert_equals: Anchored element should be at the top of anchor expected 350 but got 450
FAIL Should use the first position option with enough space below and right assert_equals: Anchored element should be at the bottom of anchor expected 199 but got 350
